Add files via upload
Model for the volcanic susceptibility around Aluto volcano based on the location of previously active vents. Probability of vent opening is calculated using a kernel denisty estimation technique, assuming a fixed bandwidth and that vent opening is a spatially non-homogenous poisson point process. The technique is based on the model described in Weller et al. 2006.